Refactor: move Visam-specific models to separate package
