+++ repondeur/zam_repondeur/visam/models/__init__.py
@@ -1,0 +1,1 @@
+from .conseil import Conseil, Formation  # noqa
