+++ repondeur/zam_repondeur/visam/views/conseils.py
@@ -6,8 +6,9 @@ from pyramid.response import Response
 from pyramid.view import view_config, view_defaults
 
 from zam_repondeur.message import Message
-from zam_repondeur.models import Chambre, Conseil, Formation
+from zam_repondeur.models import Chambre
 from zam_repondeur.services.fetch.dates import parse_date
+from zam_repondeur.visam.models import Conseil, Formation
 from zam_repondeur.visam.resources import ConseilCollection
 
 logger = logging.getLogger(__name__)
